+++ src/globalState/GlobalState.js
@@ -6,6 +6,7 @@ export const GlobalState = (props) => {
     const [user, setUser] = useState()
     const [history, setHistory] = useState()
     const [rest, setRest] = useState([])
+    console.log(user)
     const dataForm = {
         editAddressData: useForm({street:"", number:"", neighbourhood:"", city:"", state: "", complement:""}),
         editPersonalData: useForm({name:"", email:"", cpf:""}),
