+++ src/pages/ProfilePage/styled.js
@@ -3,9 +3,9 @@ import styled from "styled-components";
 export const ScreenContainer = styled.div`
 display: flex;
 flex-direction: column;
+gap: 20px;
 align-items: center;
 text-align: center;
-width: 100vw; 
 `
 export const InputsContainer = styled.div`
 display: flex;
@@ -13,4 +13,10 @@ flex-direction: column;
 width: 80vw;
 align-items: center;
 margin-bottom: 20px;
+`
+export const EnderecoContainer = styled.div`
+
+`
+export const InformacoesContainer = styled.div`
+
 `
